refactor: extract GitHubExportProvider+IssueBody (#639 slice D)
Fourth slice of #639. Moves the Markdown-body composition cluster
into a dedicated Sources/BugNarrator/Services/GitHubExportProvider+IssueBody.swift
extension file.

Moved (4 methods):
- static neutralizingUntrustedMarkdown(_:) — already internal, unchanged
- makeIssueBody(issue:session:exportFingerprint:) — widen private → internal (called by makeURLRequest in IC cluster, stays in base)
- reproductionStepReference(_:session:) — stays private (single-caller makeIssueBody moves with)
- annotatedScreenshotLines(issue:session:) — stays private (single-caller makeIssueBody moves with)

Stored-property widening:
- annotationRenderer: private let → let (used by annotatedScreenshotLines in extension file; stored properties can't live in extensions, so annotationRenderer stays declared on the actor with widened visibility)

retryConfiguration, logger stay private (minimum widening principle).

Line-count delta:
- GitHubExportProvider.swift: 519 → 358 (-161 lines)
- GitHubExportProvider+IssueBody.swift (new): 168 lines
- xcodeproj: +4 (xcodegen)

Test coverage: GitHubExportProviderTests continues to pass unchanged
(neutralizingUntrustedMarkdown test refs at :265/:271/:274 continue
to work — that method's visibility didn't change).

Closes #882. Continues #639 after PRs #875/#879/#881.
